moved clean_url and clean_urls_in_text to markdown mod
